+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 ## 0.12.30
 
+* Platform selectors for operating systems now work for Node.js tests
+  ([#742][]).
+
 * `fail()` is now typed to return `Null`, so it can be used in the same places
   as a raw `throw`.
 
 * Run Chrome in headless mode unless debugging is enabled.
+
+[#742]: https://github.com/dart-lang/test/issues/742
 
 ## 0.12.29+1